+++ lib/chef-dk/policyfile_services/install.rb
@@ -34,10 +34,12 @@ module ChefDK
       attr_reader :ui
       attr_reader :storage_config
       attr_reader :overwrite
+      attr_reader :chef_config
 
-      def initialize(policyfile: nil, ui: nil, root_dir: nil, overwrite: false)
+      def initialize(policyfile: nil, ui: nil, root_dir: nil, overwrite: false, config: nil)
         @ui = ui
         @overwrite = overwrite
+        @chef_config = config
 
         policyfile_rel_path = policyfile || "Policyfile.rb"
         policyfile_full_path = File.expand_path(policyfile_rel_path, root_dir)
@@ -66,7 +68,7 @@ module ChefDK
       end
 
       def policyfile_compiler
-        @policyfile_compiler ||= ChefDK::PolicyfileCompiler.evaluate(policyfile_content, policyfile_expanded_path, ui: ui)
+        @policyfile_compiler ||= ChefDK::PolicyfileCompiler.evaluate(policyfile_content, policyfile_expanded_path, ui: ui, chef_config: chef_config)
       end
 
       def expanded_run_list